refactor header. using our components like before
