+++ components/ui/mobile-menu.tsx
@@ -1,130 +1,213 @@
 "use client";
 
-import { useState, useRef, useEffect, SetStateAction, Dispatch } from "react";
+import { useState, useEffect } from "react";
 import Link from "next/link";
 import { Button } from "./button";
-import { Menu, X } from "lucide-react";
-import { UserResponse } from "@supabase/supabase-js";
-
-export const USER_NOT_FOUND: string = "User not found";
-
-type SupabaseUserType = UserResponse | typeof USER_NOT_FOUND;
+import { LogIn, LogOut, Menu, Settings, SquareArrowRight } from "lucide-react";
+import {
+  Sheet,
+  SheetContent,
+  SheetHeader,
+  SheetTitle,
+  SheetTrigger,
+} from "@/components/ui/sheet";
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from "@/components/ui/accordion";
+import { useTheme } from "next-themes";
+import { MoonStar, Sun } from "lucide-react";
+import { User } from "@supabase/supabase-js";
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 
 export default function MobileMenu({
+  user,
   handleSignOut,
-  supabaseUser,
 }: {
-  handleSignOut: () => Promise<never>;
-  supabaseUser: SupabaseUserType;
+  user: User | null;
+  handleSignOut: () => Promise<void>;
 }) {
   const [isOpen, setIsOpen] = useState(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-  const buttonRef = useRef<HTMLButtonElement>(null);
+  const { theme, setTheme } = useTheme();
+  const [mounted, setMounted] = useState(false);
 
   useEffect(() => {
-    const handleOutsideClick = (event: MouseEvent) => {
-      if (
-        isOpen &&
-        menuRef.current &&
-        !menuRef.current.contains(event.target as Node) &&
-        !buttonRef.current?.contains(event.target as Node)
-      ) {
-        setIsOpen(false);
-      }
-    };
+    setMounted(true);
+  }, []);
 
-    const handleEscKey = (event: KeyboardEvent) => {
-      if (isOpen && event.key === "Escape") {
-        setIsOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleOutsideClick);
-    document.addEventListener("keydown", handleEscKey);
-
-    return () => {
-      document.removeEventListener("mousedown", handleOutsideClick);
-      document.removeEventListener("keydown", handleEscKey);
-    };
-  }, [isOpen]);
-
-  const navLinks = [
-    { label: "Documentation", path: "/docs" },
-    { label: "Pricing", path: "/pricing" },
-    { label: "Priority Waitlist", path: "/priority-waitlist" },
-    { label: "FAQ", path: "/faq" },
-  ];
+  const MobileNavItem = ({
+    href,
+    onClick,
+    children,
+  }: {
+    href: string;
+    onClick: () => void;
+    children: React.ReactNode;
+  }) => (
+    <li>
+      <Link
+        href={href}
+        className="block rounded-md py-4 text-base font-medium text-foreground"
+        onClick={onClick}
+      >
+        {children}
+      </Link>
+    </li>
+  );
 
   return (
     <div className="md:hidden">
-      <button
-        ref={buttonRef}
-        className="flex items-center py-2 text-gray-700 hover:text-gray-600 focus:outline-none"
-        aria-label="Toggle mobile menu"
-        onClick={() => setIsOpen(!isOpen)}
-      >
-        <div className="relative">
-          <Menu
-            className={`h-6 w-6 transition-opacity duration-300 ease-in-out ${isOpen ? "opacity-0" : "opacity-100"}`}
-          />
-          <X
-            className={`absolute left-0 top-0 h-6 w-6 transition-opacity duration-300 ease-in-out ${isOpen ? "opacity-100" : "opacity-0"}`}
-          />
-        </div>
-      </button>
-
-      <div
-        ref={menuRef}
-        className={`absolute left-0 top-full z-20 w-full border-b border-gray-400/20 bg-gray-50 transition-all duration-300 ease-in-out ${isOpen ? "visible opacity-100" : "invisible opacity-0"} `}
-      >
-        <ul className="px-4 py-2">
-          {navLinks.map((link) => (
-            <li key={link.label}>
-              <Link
-                className="flex w-full items-center justify-center py-2 text-gray-800 hover:text-gray-600"
-                href={link.path}
-                onClick={() => setIsOpen(false)}
-              >
-                {link.label}
-              </Link>
-            </li>
-          ))}
-        </ul>
-        <div className="space-y-2 px-4 pb-4">
-          {supabaseUser === USER_NOT_FOUND ? (
-            <div className="flex items-center justify-center space-x-2.5">
-              <Button asChild className="w-full rounded-full">
-                <Link onClick={() => setIsOpen(false)} href={"/signin"}>
-                  Sign in
-                </Link>
-              </Button>
-              <Button asChild className="w-full" variant="outline">
-                <Link onClick={() => setIsOpen(false)} href={"/signup"}>
-                  Sign up
-                </Link>
-              </Button>
+      <Sheet open={isOpen} onOpenChange={setIsOpen}>
+        <SheetTrigger asChild>
+          <Button
+            variant="ghost"
+            size="icon"
+            className="rounded-md"
+            aria-label="Open menu"
+          >
+            <Menu className="h-6 w-6" />
+          </Button>
+        </SheetTrigger>
+        <SheetContent
+          side="right"
+          className="flex w-[300px] flex-col justify-between sm:w-[400px]"
+        >
+          <div>
+            <SheetHeader className="mb-4">
+              <SheetTitle>PearAI Menu</SheetTitle>
+            </SheetHeader>
+            <nav aria-label="Mobile menu">
+              <ul className="space-y-1">
+                <div className="mb-4 space-y-4">
+                  {user ? (
+                    <>
+                      <div className="flex hidden items-center space-x-2">
+                        <Avatar className="h-8 w-8">
+                          <AvatarImage
+                            src={user.user_metadata.avatar_url}
+                            alt={user.user_metadata.full_name || "User avatar"}
+                          />
+                          <AvatarFallback>
+                            {user.email?.[0].toUpperCase() || "U"}
+                          </AvatarFallback>
+                        </Avatar>
+                        <span className="text-sm font-medium">
+                          {user.user_metadata.full_name || user.email}
+                        </span>
+                      </div>
+                      <Link href="/dashboard" onClick={() => setIsOpen(false)}>
+                        <Button
+                          variant="outline"
+                          className="w-full justify-start"
+                        >
+                          <Settings className="mr-2 h-4 w-4" />
+                          Dashboard
+                        </Button>
+                      </Link>
+                      <Button
+                        variant="outline"
+                        className="w-full justify-start"
+                        onClick={() => {
+                          handleSignOut();
+                          setIsOpen(false);
+                        }}
+                      >
+                        <LogOut className="mr-2 h-4 w-4" />
+                        Sign out
+                      </Button>
+                    </>
+                  ) : (
+                    <>
+                      <Link href="/signin" onClick={() => setIsOpen(false)}>
+                        <Button
+                          variant="outline"
+                          className="w-full justify-start"
+                        >
+                          <LogIn className="mr-2 h-4 w-4" />
+                          Sign in
+                        </Button>
+                      </Link>
+                      <Link href="/signup" onClick={() => setIsOpen(false)}>
+                        <Button
+                          variant="outline"
+                          className="mt-4 w-full justify-start"
+                        >
+                          <SquareArrowRight className="mr-2 h-4 w-4" />
+                          Try PearAI
+                        </Button>
+                      </Link>
+                    </>
+                  )}
+                </div>
+                <MobileNavItem href="/" onClick={() => setIsOpen(false)}>
+                  Home
+                </MobileNavItem>
+                <Accordion type="single" collapsible className="w-full">
+                  <AccordionItem value="resources">
+                    <AccordionTrigger>Resources</AccordionTrigger>
+                    <AccordionContent>
+                      <ul className="ml-4 space-y-1">
+                        <MobileNavItem
+                          href="/about"
+                          onClick={() => setIsOpen(false)}
+                        >
+                          About
+                        </MobileNavItem>
+                        <MobileNavItem
+                          href="/docs"
+                          onClick={() => setIsOpen(false)}
+                        >
+                          Documentation
+                        </MobileNavItem>
+                        <MobileNavItem
+                          href="/faq"
+                          onClick={() => setIsOpen(false)}
+                        >
+                          FAQ
+                        </MobileNavItem>
+                        <MobileNavItem
+                          href="/changelog"
+                          onClick={() => setIsOpen(false)}
+                        >
+                          Changelog
+                        </MobileNavItem>
+                      </ul>
+                    </AccordionContent>
+                  </AccordionItem>
+                </Accordion>
+                <MobileNavItem href="/pricing" onClick={() => setIsOpen(false)}>
+                  Pricing
+                </MobileNavItem>
+              </ul>
+            </nav>
+          </div>
+          <div className="width-full space-y-4 pb-6">
+            <div className="width-full">
+              {mounted ? (
+                <Button
+                  variant="outline"
+                  className="w-full justify-center"
+                  onClick={() => setTheme(theme === "light" ? "dark" : "light")}
+                >
+                  {theme === "light" ? (
+                    <>
+                      <Sun strokeWidth={1} className="h-5 w-5" />
+                      Light
+                    </>
+                  ) : (
+                    <>
+                      <MoonStar strokeWidth={1} className="h-5 w-5" />
+                      Dark
+                    </>
+                  )}
+                </Button>
+              ) : null}
             </div>
-          ) : (
-            <div className="flex items-center justify-center space-x-2.5">
-              <Button asChild className="w-full rounded-full">
-                <Link onClick={() => setIsOpen(false)} href={"/dashboard"}>
-                  Dashboard
-                </Link>
-              </Button>
-              <Button
-                onClick={() => setIsOpen(false)}
-                asChild
-                variant="outline"
-                className="w-full"
-              >
-                <form action={handleSignOut}>
-                  <button className="w-full">Sign out</button>
-                </form>
-              </Button>
-            </div>
-          )}
-        </div>
-      </div>
+          </div>
+        </SheetContent>
+      </Sheet>
     </div>
   );
 }
